requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-admin-honeypot==1.1.0
django-audittrail==0.2.1
django-environ==0.4.5
django-honeypot==0.9.0
gunicorn==20.0.4
idna==2.10
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tdqm==0.0.1
tqdm==4.56.0
urllib3==1.26.2
